Implement event trigger association and settiings endpoints

+++ src/main/java/com/czertainly/core/service/impl/NotificationInstanceServiceImpl.java
@@ -14,14 +14,15 @@ import com.czertainly.api.model.core.notification.AttributeMappingDto;
 import com.czertainly.api.model.core.notification.NotificationInstanceDto;
 import com.czertainly.api.model.core.notification.NotificationInstanceRequestDto;
 import com.czertainly.api.model.core.notification.NotificationInstanceUpdateRequestDto;
-import com.czertainly.api.model.core.settings.NotificationSettingsDto;
 import com.czertainly.api.model.core.settings.SettingsSection;
 import com.czertainly.core.attribute.engine.AttributeEngine;
 import com.czertainly.core.dao.entity.Connector;
 import com.czertainly.core.dao.entity.notifications.NotificationInstanceMappedAttributes;
 import com.czertainly.core.dao.entity.notifications.NotificationInstanceReference;
+import com.czertainly.core.dao.entity.workflows.Execution;
 import com.czertainly.core.dao.repository.notifications.NotificationInstanceMappedAttributeRepository;
 import com.czertainly.core.dao.repository.notifications.NotificationInstanceReferenceRepository;
+import com.czertainly.core.dao.repository.notifications.NotificationProfileVersionRepository;
 import com.czertainly.core.model.auth.ResourceAction;
 import com.czertainly.core.security.authz.ExternalAuthorization;
 import com.czertainly.core.security.authz.SecuredUUID;
@@ -48,6 +49,8 @@ public class NotificationInstanceServiceImpl implements NotificationInstanceServ
 
     private NotificationInstanceReferenceRepository notificationInstanceReferenceRepository;
     private NotificationInstanceMappedAttributeRepository notificationInstanceMappedAttributeRepository;
+    private NotificationProfileVersionRepository notificationProfileVersionRepository;
+
     private ConnectorService connectorService;
     private CredentialService credentialService;
     private NotificationInstanceApiClient notificationInstanceApiClient;
@@ -67,6 +70,11 @@ public class NotificationInstanceServiceImpl implements NotificationInstanceServ
     @Autowired
     public void setNotificationInstanceMappedAttributeRepository(NotificationInstanceMappedAttributeRepository notificationInstanceMappedAttributeRepository) {
         this.notificationInstanceMappedAttributeRepository = notificationInstanceMappedAttributeRepository;
+    }
+
+    @Autowired
+    public void setNotificationProfileVersionRepository(NotificationProfileVersionRepository notificationProfileVersionRepository) {
+        this.notificationProfileVersionRepository = notificationProfileVersionRepository;
     }
 
     @Autowired
@@ -263,24 +271,13 @@ public class NotificationInstanceServiceImpl implements NotificationInstanceServ
         } else {
             logger.debug("Deleting notification without connector: {}", notificationInstanceRef);
         }
-        notificationInstanceReferenceRepository.delete(notificationInstanceRef);
 
-        // check notifications settings and remove deleted instance if used
-        NotificationSettingsDto notificationsSettings = SettingsCache.getSettings(SettingsSection.NOTIFICATIONS);
-        if (notificationsSettings != null) {
-            boolean updated = false;
-            for (NotificationType notificationType : NotificationType.values()) {
-                String notificationInstanceUuid = notificationsSettings.getNotificationsMapping().get(notificationType);
-                if (notificationInstanceUuid != null && UUID.fromString(notificationInstanceUuid).equals(notificationInstanceRef.getUuid())) {
-                    updated = true;
-                    notificationsSettings.getNotificationsMapping().remove(notificationType);
-                }
-            }
-
-            if (updated) {
-                logger.debug("Updating notifications settings. Removing deleted notification instance");
-                settingService.updateNotificationSettings(notificationsSettings);
-            }
+        // check notification profiles referencing notification instance
+        long referencesCount = notificationProfileVersionRepository.countByNotificationInstanceRefUuid(notificationInstanceRef.getUuid());
+        if (referencesCount > 0) {
+            throw new ValidationException("Cannot delete notification instance. %d notification profile version(s) are referencing this notification instance".formatted(referencesCount));
         }
+
+        notificationInstanceReferenceRepository.delete(notificationInstanceRef);
     }
 }
